reafactor: types and settings form
