+++ rslang/src/app/modules/learning-mode/components/settings/settings.component.scss
@@ -9,4 +9,8 @@
   .form-setting {
     margin: 25px;
   }
+
+  .translate {
+    margin-top: 25px;
+  }
 }
