--- rslang/src/app/shared/services/types.ts
@@ -1,2 +1,0 @@
-export type Group = 0|1|2|3|4|5;
-export type Page = 0|1|2|3|4|5|6|7|8|9|10|11|12|13|14|15|16|17|18|19|20|21|22|23|24|25|26|27|28|29;
